Updated IEmbeddedMessage in events
