+++ src/index.ts
@@ -3,6 +3,7 @@ export * from './users';
 export * from './inapp';
 export * from './events';
 export * from './commerce';
+export * from './types';
 export * from './embedded';
 export * from './components/card';
 export * from './components/banner';
